config updates, moved map style to its own file
