+++ src/styles/MapBackground.module.css
@@ -11,11 +11,32 @@
 	/* visibility: hidden; */
 	background-color: hsla(108, 45%, 96%, 1);
 	background-image: url("../img/map-background-noise-light.png");
-	/* opacity: 0.7; */
+	opacity: 0.7;
 	position: fixed;
 	height: 100%;
 	width: 100%;
 	z-index: 1;
 	backdrop-filter: blur(2.5px);
 	transition: all 4s ease-in-out;
+	transition-delay: 0.333s;
+}
+
+.overlay-nightmode {
+	background-color: hsla(218, 15%, 05%, 0.56);
+	opacity: 0.4;
+	backdrop-filter: blur(4px);
+}
+
+.overlay-rma-android {
+	background-color: hsla(0, 0%, 100%, 0.49);
+	opacity: 0.5;
+	backdrop-filter: blur(1px);
+	/* transition: background-image 4s ease-in-out; */
+	transition-delay: 0.333s;
+}
+
+.overlay-qing-dynasty-map {
+	background-color: hsla(0, 0%, 24%, 0.6);
+	opacity: 0.4;
+	backdrop-filter: blur(2px);
 }
